requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.9.1
torchmetrics>=0.11.0

# --------- hydra --------- #
hydra-core>=1.3.1
hydra-colorlog>=1.2.0
hydra-optuna-sweeper>=1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils~=1.0.4
pre-commit      # hooks for applying linters on commit
rich~=12.5.1
pytest~=7.1.2
# sh            # for running bash commands in some tests (linux/macos only)

omegaconf~=2.2.3
torchinfo~=1.7.1
numpy~=1.23.2
pillow~=9.2.0
packaging~=21.3
sh~=1.14.3
setuptools~=63.4.1

opencv-python~=4.5.4.60
matplotlib~=3.5.3
kornia~=0.6.9
qrcode~=7.4.2